refactor: consolidate util files
